+++ packages/editor/js/extensions/libs/block-card/inner-blocks/types/index.js
@@ -55,6 +55,7 @@ export type InnerBlocksProps = {
 	onChange: THandleOnChangeAttributes,
 	maxItems: number,
 	contextValue: Object,
+	doingSwitchToInner: () => void,
 	setCurrentBlock: (block: InnerBlockType) => void,
 	setBlockClientInners: ({
 		clientId: string,