+++ packages/editor/js/extensions/libs/block-composite/components/categorized-items.js
@@ -3,8 +3,8 @@
 /**
  * External dependencies
  */
-import { __, sprintf } from '@wordpress/i18n';
 import type { MixedElement } from 'react';
+import { __, sprintf } from '@wordpress/i18n';
 import { Icon as WordPressIconComponent } from '@wordpress/components';
 
 /**
@@ -19,7 +19,6 @@ import { Flex, Grid, Tooltip } from '@blockera/controls';
  */
 import { getTooltipStyle } from '../utils';
 import type { TCategorizedItemsProps } from '../types';
-import { isInnerBlock } from '../../../components/utils';
 import type { InnerBlockModel } from '../../block-card/inner-blocks/types';
 import type { TStates, StateTypes } from '../../block-card/block-states/types';
 
@@ -37,7 +36,7 @@ export const CategorizedItems = ({
 	getBlockInners,
 	limited = false,
 	setCurrentBlock,
-	supportsInnerBlocks,
+	doingSwitchToInner,
 	setBlockClientInners,
 }: TCategorizedItemsProps): MixedElement => {
 	// Filtering items while not exists on received category ...
@@ -68,6 +67,8 @@ export const CategorizedItems = ({
 
 		if (name) {
 			const inners = getBlockInners(clientId);
+
+			doingSwitchToInner();
 
 			setBlockClientInners({
 				clientId,
@@ -293,12 +294,7 @@ export const CategorizedItems = ({
 									}
 									className={classNames(
 										'blockera-feature-type',
-										'is-item',
-										{
-											'not-allowed':
-												false === supportsInnerBlocks &&
-												isInnerBlock(id),
-										}
+										'is-item'
 									)}
 								>
 									{icon && (