+++ packages/editor/js/extensions/libs/block-composite/components/preview.js
@@ -3,19 +3,22 @@
 /**
  * External dependencies
  */
+import { select } from '@wordpress/data';
 import type { MixedElement } from 'react';
+import { useCallback } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { Inserter, Categories } from './';
-import type { TPreviewProps } from '../types';
-import { useBlockStates } from '../../block-card/block-states/hooks';
-import StatesManager from '../../block-card/block-states/components/states-manager';
 import {
 	InnerBlocksExtension,
 	useInnerBlocks,
 } from '../../block-card/inner-blocks';
+import { Inserter, Categories } from './';
+import type { TPreviewProps } from '../types';
+import { useBlockStates } from '../../block-card/block-states/hooks';
+import type { TStates, StateTypes } from '../../block-card/block-states/types';
+import StatesManager from '../../block-card/block-states/components/states-manager';
 
 // the instance of in-memory cache.
 const deleteCacheData: Object = new Map();
@@ -76,18 +79,50 @@ export const Preview = ({
 		currentInnerBlockState,
 	});
 
+	const doingSwitchToInner = useCallback(() => {
+		const { getState, getInnerState } = select('blockera/editor');
+		const {
+			settings: { supportsInnerBlocks },
+		} = getState(currentState) ||
+			getInnerState(currentInnerBlockState) || {
+				settings: { supportsInnerBlocks: true },
+			};
+
+		if (false === supportsInnerBlocks) {
+			const newStates: {
+				[key: TStates]: StateTypes,
+				// $FlowFixMe
+			} = {
+				...states,
+				normal: {
+					...states.normal,
+					isSelected: true,
+					selectable: true,
+				},
+			};
+
+			// Reset isSelected flag for all other states
+			Object.keys(newStates).forEach((stateName) => {
+				if (stateName !== 'normal') {
+					// $FlowFixMe
+					newStates[stateName].isSelected = false;
+				}
+			});
+
+			handleOnChange(newStates);
+		}
+	}, [states, handleOnChange, currentState, currentInnerBlockState]);
+
 	return (
 		<StatesManager
 			states={states}
 			onDelete={onDelete}
-			currentState={currentState}
 			overrideItem={overrideItem}
 			defaultStates={defaultStates}
 			preparedStates={preparedStates}
 			handleOnChange={handleOnChange}
 			deleteCacheData={deleteCacheData}
 			contextValue={blockStatesContextValue}
-			currentInnerBlockState={currentInnerBlockState}
 			defaultRepeaterItemValue={defaultRepeaterItemValue}
 			maxItems={Object.keys(preparedStates).length + (maxItems || 0)}
 			getDynamicDefaultRepeaterItem={getDynamicDefaultRepeaterItem}
@@ -98,13 +133,7 @@ export const Preview = ({
 							...props,
 							maxItems:
 								maxItems + Object.keys(preparedStates).length,
-							currentState,
-							currentInnerBlockState,
-							AvailableBlocks: ({
-								supportsInnerBlocks,
-							}: {
-								supportsInnerBlocks: boolean,
-							}) => (
+							AvailableBlocks: () => (
 								<Categories
 									blocks={blocks}
 									elements={elements}
@@ -114,8 +143,8 @@ export const Preview = ({
 									setBlockState={handleOnChange}
 									getBlockInners={getBlockInners}
 									setCurrentBlock={setCurrentBlock}
+									doingSwitchToInner={doingSwitchToInner}
 									getBlockStates={() => calculatedStates}
-									supportsInnerBlocks={supportsInnerBlocks}
 									setBlockClientInners={setBlockClientInners}
 								/>
 							),
@@ -129,10 +158,12 @@ export const Preview = ({
 					{...{
 						...innerBlocksProps,
 						maxItems,
-						setCurrentBlock,
-						setBlockClientInners,
 						currentState,
+						setCurrentBlock,
 						currentBreakpoint,
+						doingSwitchToInner,
+						setBlockClientInners,
+						currentInnerBlockState,
 						contextValue: innerBlocksContextValue,
 					}}
 					block={block}