+++ packages/features-library/icon/src/functions.php
@@ -12,6 +12,15 @@ if ( ! function_exists('blockera_block_has_icon')) {
 		$has_class = str_contains($block['attrs']['className'] ?? '', 'blockera-has-icon');
 		$has_icon  = isset($block['attrs']['blockeraIcon']['value']) && ! empty($block['attrs']['blockeraIcon']['value']);
 
+		// Try to check if the block has an icon in the inner blocks.
+		if (! $has_class && ! $has_icon) {
+			foreach ($block['innerBlocks'] as $innerBlock) {
+				if (blockera_block_has_icon($innerBlock)) {
+					return true;
+				}
+			}
+		}
+
 		return $has_class && $has_icon;
 	}
 }