+++ tests/fixtures/block-post-author-biography/setup.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * Setup for the post featured image block.
+ * 
+ * @var string $post_content The post content.
+ * @var BlockeraTest $this The test instance.
+ * @var string $designName The design name.
+ */
+
+update_user_meta( 1, 'description', 'This is a sample biography for the test WordPress user.' );